requirements.txt
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
markdown>=3.5.0
weasyprint>=60.0
tqdm>=4.66.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
greenlet>=3.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0